% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigopt_api_wrapper.R
\name{fetch_experiment}
\alias{fetch_experiment}
\title{Fetch an experiment}
\usage{
fetch_experiment(experiment_id, body = NULL)
}
\arguments{
\item{experiment_id}{the id of an experiment to fetch}

\item{body}{Url params of GET request}
}
\value{
SigOpt experiment with id experiment_id
}
\description{
Fetch an experiment
}
\examples{
fetch_experiment(5807)
}

