% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{franke}
\alias{franke}
\title{Franke function - http://www.sfu.ca/~ssurjano/franke2d.html}
\usage{
franke(x, y)
}
\arguments{
\item{x}{First dimension}

\item{y}{Second dimension}
}
\value{
The franke function evaluated at x,y
}
\description{
Franke function - http://www.sfu.ca/~ssurjano/franke2d.html
}
\examples{
franke(0,1)
}
