% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigopt_api_wrapper.R
\name{sigopt_api_token}
\alias{sigopt_api_token}
\title{Get the SigOpt API token from the SIGOPT_API_TOKEN environment variable or user input}
\usage{
sigopt_api_token(force = FALSE)
}
\arguments{
\item{force}{force entry of SigOpt API token, even if present}
}
\value{
SigOpt API token
}
\description{
Get the SigOpt API token from the SIGOPT_API_TOKEN environment variable or user input
}
