% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigopt_api_wrapper.R
\name{sigopt_api_url}
\alias{sigopt_api_url}
\title{Get the SigOpt API url from the SIGOPT_API_URL environment variable or use default
Most users will be ok with the default value}
\usage{
sigopt_api_url()
}
\value{
Base url for SigOpt API requests
}
\description{
Get the SigOpt API url from the SIGOPT_API_URL environment variable or use default
Most users will be ok with the default value
}
\seealso{
\code{\link{sigopt_GET}} and \code{\link{sigopt_POST}}, which perform the HTTP requests
}
