% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigopt_api_wrapper.R
\name{sigopt_api_user_agent}
\alias{sigopt_api_user_agent}
\title{User agent for current version of SigOpt R API Client}
\usage{
sigopt_api_user_agent()
}
\value{
User agent
}
\description{
User agent for current version of SigOpt R API Client
}
\seealso{
\code{\link{sigopt_GET}} and \code{\link{sigopt_POST}}, which perform the HTTP requests
}
