% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigopt_api_wrapper.R
\name{sigopt_auth}
\alias{sigopt_auth}
\title{Create authentication for SigOpt API, using HTTP Basic Auth}
\usage{
sigopt_auth(api_token = sigopt_api_token())
}
\arguments{
\item{api_token}{SigOpt API token}
}
\value{
http basic authentiation with api_token as username and no password
}
\description{
Create authentication for SigOpt API, using HTTP Basic Auth
}
