% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigopt_api_wrapper.R
\name{sigopt_check}
\alias{sigopt_check}
\title{Check content returned by the SigOpt API}
\usage{
sigopt_check(req)
}
\arguments{
\item{req}{result of request to the SigOpt API}
}
\value{
invisible(), stops if there is an error status code
}
\description{
Check content returned by the SigOpt API
}
\seealso{
\code{\link{sigopt_GET}} and \code{\link{sigopt_POST}}, which call this function interally
}
