% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigopt_api_wrapper.R
\name{sigopt_GET}
\alias{sigopt_GET}
\title{GET request to SigOpt API path, with json encoded body}
\usage{
sigopt_GET(path, query = NULL, api_token = sigopt_api_token())
}
\arguments{
\item{path}{path of SigOpt API url to POST to}

\item{query}{list of query parameters to be url-encoded}

\item{api_token}{SigOpt api token}
}
\value{
result of request to SigOpt API, needs to be JSON decoded
}
\description{
GET request to SigOpt API path, with json encoded body
}
\seealso{
\code{\link{sigopt_parse}}, which parses the result of this function
}
