% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigopt_api_wrapper.R
\name{sigopt_has_api_token}
\alias{sigopt_has_api_token}
\title{Whether or not the SigOpt API token is properly configured}
\usage{
sigopt_has_api_token()
}
\value{
TRUE or FALSE
}
\description{
Whether or not the SigOpt API token is properly configured
}
\examples{
sigopt_has_api_token()
}
\seealso{
\code{\link{sigopt_api_token}}, which this function uses to get the SigOpt API token
}

