% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigopt_api_wrapper.R
\name{sigopt_parse}
\alias{sigopt_parse}
\title{Parse content returned by the SigOpt API}
\usage{
sigopt_parse(req)
}
\arguments{
\item{req}{result of request to the SigOpt API}
}
\value{
json decoding of request object
}
\description{
Parse content returned by the SigOpt API
}
\examples{
sigopt_parse(GET("https://api.sigopt.com", path="v1/experiments", sigopt_auth(sigopt_api_token())))
sigopt_parse(POST("https://api.sigopt.com", path="v1/experiments", sigopt_auth(sigopt_api_token())))
}
\seealso{
\code{\link{sigopt_GET}} and \code{\link{sigopt_POST}}, which call this function interally
}

