% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigopt_api_wrapper.R
\name{sigopt_POST}
\alias{sigopt_POST}
\title{POST request to SigOpt API path, with json encoded body}
\usage{
sigopt_POST(path, body, api_token = sigopt_api_token())
}
\arguments{
\item{path}{path of SigOpt API url to POST to}

\item{body}{POST body, will be json-encoded}

\item{api_token}{SigOpt api token}
}
\value{
result of request to SigOpt API, needs to be JSON decoded
}
\description{
POST request to SigOpt API path, with json encoded body
}
\seealso{
\code{\link{sigopt_parse}}, which parses the result of this function
}
