% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigopt_api_wrapper.R
\name{sigopt_parse}
\alias{sigopt_parse}
\title{Parse content returned by the SigOpt API}
\usage{
sigopt_parse(req)
}
\arguments{
\item{req}{result of request to the SigOpt API}
}
\value{
json decoding of request object
}
\description{
Parse content returned by the SigOpt API
}
\examples{
sigopt_parse(sigopt_GET("v1/experiments", null))
sigopt_parse(sigopt_POST("v1/experiments/5807/suggestions", null))
}
\seealso{
\code{\link{sigopt_GET}} and \code{\link{sigopt_POST}}, which call this function interally
}

